fix: stuck at no text-content post
